models:identification;database:files_db - move "count" models to database
