+++ acacore/models/identification.py
@@ -27,18 +27,3 @@ class Identification(ACABase):
             raise ValueError(f"PUID missing for signature {signature}.")
 
         return data
-
-
-class SignatureCount(ACABase):
-    """Signature count datamodel."""
-
-    puid: Optional[str]
-    signature: Optional[str]
-    count: Optional[int]
-
-
-class ChecksumCount(ACABase):
-    """Signature count datamodel."""
-
-    checksum: str
-    count: int
